fix nosecboot
